Fix http://47.116.78.131:10083/issues/2709

+++ pkg/category/delete.go
@@ -4,9 +4,11 @@ import (
 	"context"
 	"fmt"
 
+	articlemwcli "github.com/NpoolPlatform/cms-middleware/pkg/client/article"
 	categorymwcli "github.com/NpoolPlatform/cms-middleware/pkg/client/category"
 	cruder "github.com/NpoolPlatform/libent-cruder/pkg/cruder"
 	basetypes "github.com/NpoolPlatform/message/npool/basetypes/v1"
+	articlemwpb "github.com/NpoolPlatform/message/npool/cms/mw/v1/article"
 	npool "github.com/NpoolPlatform/message/npool/cms/mw/v1/category"
 )
 
@@ -21,6 +23,17 @@ func (h *Handler) DeleteCategory(ctx context.Context) (*npool.Category, error) {
 	}
 	if !exist {
 		return nil, fmt.Errorf("invalid category")
+	}
+
+	exist, err = articlemwcli.ExistArticleConds(ctx, &articlemwpb.Conds{
+		AppID:      &basetypes.StringVal{Op: cruder.EQ, Value: *h.AppID},
+		CategoryID: &basetypes.StringVal{Op: cruder.EQ, Value: *h.EntID},
+	})
+	if err != nil {
+		return nil, err
+	}
+	if exist {
+		return nil, fmt.Errorf("exist article")
 	}
 
 	return categorymwcli.DeleteCategory(ctx, *h.ID)
